requirements.txt
# Vector Database
qdrant-client

# OpenAI API
openai

# Utilities
tiktoken
tqdm
python-dotenv

# Visualization
matplotlib
